+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Allow to use extended capabilities and scene props
+* Allow to send and receive custom proto packet
+
 ## 2022-04-26 v1.3.1
 * Fix type definition for `onSession` method